예외 상황 발생시의 처리 로직을 각 메소드마다 구현

+++ src/main/java/com/hancom/board/controller/BoardController.java
@@ -30,6 +30,7 @@ public class BoardController {
 	private BoardService boardService;
 	
 	//목록
+	//게시판에 게시물이 한 개도 없더라도 원하는 정보를 찾지 못하는 오류 상황은 아니다. 따라서 NPE 예외 처리를 작성하지 않았고 빈 리스트를 리턴해준다.
 	@ApiOperation(value = "목록", notes = "게시판의 전체 목록을 조회한다. 해당 API는 게시물 번호, 제목, 작성자, 최초 작성일의 정보만 제공한다.")
 	@GetMapping("/board")
 	public List<BoardOutputDTO> get(){
@@ -39,10 +40,10 @@ public class BoardController {
 		
 		//데이터베이스에 저장되어있는 모든 레코드를 boardList에 가져온 뒤 그 리스트의 각 요소들의 본문내용과 최신 수정일 필드를 제외하고 새로운 리스트를 만들어 반환해준다. 
 		boardList.forEach(e -> {
-			boardOutputDTOList.add(BoardOutputDTO.builder().number(e.getNumber()).title(e.getTitle()).
-					author(e.getAuthor()).createdDate(e.getCreatedDate()).build());
+			boardOutputDTOList.add(BoardOutputDTO.builder().number(e.getNumber()).title(e.getTitle())
+					.author(e.getAuthor()).createdDate(e.getCreatedDate()).build());
 		});
-		
+			
 		return boardOutputDTOList;
 	}
 	
@@ -53,6 +54,7 @@ public class BoardController {
 		
 		Board board = boardService.get(number);
 		
+		//원하는 번호의 게시물이 없을 경우에는 RuntimeException을 발생시킨다.
 		if(board == null) {
 			throw new RuntimeException(number + "번 게시물을 찾지 못했습니다.");
 		}
@@ -65,12 +67,17 @@ public class BoardController {
 	@PostMapping("/board")
 	public Board save(@RequestBody BoardInputDTO boardDTO) {
 		
+		//필수 작성 항목인 제목이나 작성자가 기입되어있지 않을 경우 RuntimeException을 발생시킨다.
+		if(boardDTO.getTitle() == null || boardDTO.getAuthor() == null) {
+			throw new RuntimeException("필수 작성 항목을 채우지않아 게시물을 저장할 수 없습니다.");
+		}
+		
 		//해당 포맷에 맞춰 최초 작성일을 저장한다.
 		SimpleDateFormat createdSimpleDate = new SimpleDateFormat("yyyy-MM-dd hh:mm:ss");
 		
-		//BoardDTO에서 Board 모델로 변경하는 작업 
-		Board board = Board.builder().title(boardDTO.getTitle()).author(boardDTO.getAuthor()).content(boardDTO.getContet()).
-				createdDate(createdSimpleDate.format(new Date()).toString()).build();
+		//BoardDTO에서 Board 모델로 변경하는 작업
+		Board board = Board.builder().title(boardDTO.getTitle()).author(boardDTO.getAuthor()).content(boardDTO.getContent())
+				.createdDate(createdSimpleDate.format(new Date()).toString()).build();
 		boardService.save(board);
 		
 		return board;
@@ -81,13 +88,14 @@ public class BoardController {
 	//그러나 이러한 경우에는 유저가 임의로 변경하면 안되는 정보(번호, 최초 작성일, 최근 수정일)를 변경하여 보낼 수도 있다.
 	//따라서 이러한 문제를 해결하기 위해 해당 메소드에서 먼저 게시물의 번호가 동일한 객체를 가져온 뒤 변경할 정보들만 변경해주는 작업을 한 뒤 데이터베이스에 저장해주는 방식으로 구현하였다.
 	//(프론트엔드 단에서 미리 방지해 줄 수도 있겠지만 백엔드는 프론트엔드에서 어떤 정보가 올지 모르기에 모든 상황을 고려해서 로직을 작성해야한다고 배웠기에 아래와 같이 로직을 구성하였다.)
-	@ApiOperation(value = "수정", notes = "게시판에 존재하던 게시물의 내용을 수정한다. 해당 API는 유저가 보내지 말아야할 정보(게시물 번호, 최초 작성일, 최근 수정일)를 보내더라도"
+	@ApiOperation(value = "수정", notes = "게시판에 존재하던 게시물의 내용을 수정한다. 해당 API는 유저가 보내지 말아야할 정보(게시물 번호, 최초 작성일, 최근 수정일)를 보내더라도 "
 			+ "백엔드 단에서 수정 가능한 정보만 변경하여 데이터베이스에 저장한다.")
 	@PutMapping("/board")
 	public Board update(@RequestBody Board board) {
 		
 		Board originBoard = boardService.get(board.getNumber());
 		
+		//수정하고자 하는 번호의 게시물이 없을 경우에는 RuntimeException을 발생시킨다.
 		if(originBoard == null) {
 			throw new RuntimeException(board.getNumber() + "번 게시물을 찾지 못했습니다.");
 		}
@@ -109,6 +117,11 @@ public class BoardController {
 	@ApiOperation(value = "삭제", notes = "게시판의 특정 게시물을 삭제한다. 해당 API는 특정 게시물 삭제가 잘 되었는지 결과를 리턴해준다.")
 	@DeleteMapping("/board/{number}")
 	public String delete(@PathVariable int number) {
+		
+		//삭제하고자 하는 번호의 게시물이 없을 경우에는 RuntimeException을 발생시킨다.
+		if(boardService.get(number) == null) {
+			throw new RuntimeException(number + "번 게시물을 삭제할 수 없습니다.");
+		}
 		
 		boardService.delete(number);
